Improved documentation, constructor and mappings
More documentation has been added to the MultipleChoiceQuestion, TestMulipleChoiceQuestion and TestMultipleChoiceQuestionJpa classes

Additional tests have been added

Added a parameter to the constructor of MultipleChoiceQuestion that takes a collection to be used for choices.

A mapping has been added and the previous mapping has been changed in MultipleChoiceQuestion class to be used easier in the future.

The first mapping choicesID maps a unique ID to the string representation of each choice.

The second mapping stores the answers given for each choice by their IDs